refactor: changed packet folder structure to be more streamlined
